Fundamental changes on TaxonPath and taxon handling have been made.
Transformation of LOM resources derived from LOM taxons and taxonPaths
cannot be generic. It requires a given taxonomy/ontology expressing the
classification of ressources for a given instituion or group of institutions.
The method implemented here agregates an external IRI root bearing the
classification and the Id tag of the LAST taxon of each taxonPathes.
(The taxonPath is known in the graph of the ontology which allows to
consider the LAST taxon only)
Taxons are linked to ontology clases or individuals while keywords
are linked to litterals.
Besides the IRI it is also necessary to qualify the taxonPath SOURCE
classification, to identify to provide the qualifier string associated
with each source.
As an example an ontology build on 4 classifications at ENS de Lyon is
used here.